Dockerized thompsonbot & reorganized files
