+++ config.json
@@ -1,0 +1,1 @@
+{"channels": {"sandbox": 1022494749560684554}, "tracking": false}
